pyproject.toml
[project]
name = "shadow-check"
version = "1.0.0"
requires-python = ">=3.13"
dependencies = [
    "playwright==1.58.0",
    "python-dotenv==1.2.2",
    "httpx==0.28.1"
]